[Behat][Checkout] Raw moving complete steps
